Schema: [RFK] moved hardwaredb to separate package
Our hardware database will not be part of the Artemis OSS
release. Therefore, I moved it from the Schema package which will be
part of the release to the new HardwareDB package which will not be
published. The HardwareDB package will also be responsible for filling
the TestrunDB HostFeature table.