[FEAT] table for denied hosts
Some queues contains tests that do not work on certain hosts. As a
convenience for the user we want to prevent jobs from such a queue to be
scheduled on the broken host without the user having to add
"requested_feature" to every test request. To keep the information on
what queue denied what host we need a new database table. This patch add
the necessary changed to the schema.