enhance pool hosts
When a pool host is scheduled some part of the MCP (e.g. producer,
resetter) may replace the host entry in the testrun_scheduling returned
by scheduler. We don't want to loose the pool association on the
replacement host and therefore need a connection back to the original
pool host. This patch provides this-