Add setter functionality to pool_count
pool_count used to be a getter for the sum of used and unused elements
of a pool host.
It's obviously also necessary to update this value. Therefore, I added
setter functionality to this function. Since there is no field for the
sum, this works by calculating the number of used pool elements,
substracting it from the requested number and putting the result in as
new pool_free.