[FIX] updating pool hosts for marking jobs as running and finished
When a job is marked as running of finished, its associated host needs
to be updated. For pool hosts, this means the pool_free counter needs to
be updated. To make this update correctly, we need to reread the current
value from the database into the DBIC object and then execute the update
in a transaction.