[FIX] host_free on finished testrun for pool hosts
A pool host can have a pool_free below zero when the user reduced
the pool_count below the number of currently used pool elements. In this
case, when a testrun on one of the pool elements is finished, we can not
always set the pool host to free. Instead, we need to check whether
pool_free is above zero.