harmonize types to fit constraints
There was a mixtures to INT/TINYINT/SMALLINT
which not always fit together when referring
to their foreign keys in other tables.

This is now harmonized. If in doubt, the bigger
datatype was used, i.e., TINY->SMALL, SMALL->INTEGER.

And for the records, TINYINT should become SMALLINT
to easier support other databases like PostgreSQL.