fix transaction handling in job state toggle
When a testrun goes into state 'running' then the
respective host it is scheduled on is set to 'free=0'.

Both steps belong together, atomically.

However, the setting of the testrun state happened
outside the transaction in which the host free setting
happened.

This means the setting of the host free could be
rolled back while the testrun is already running.

This means on the next scheduling cycle the host
is found again as being free tand the next testrun
is scheduled on it.

Now we pull the testrun state setting into the
transaction and also reshuffle the instructions a
bit to first set the host free=0, and then the
testrun=running, so that we later can do that
in 2 steps again when multiple schedulers might
run, so the host is locked as soon as possible.