double check conditions when setting host free=1
This is mostly a work-around for a situation that
should not happen since the previous commit when a
host was double-booked.

That condition now at least does not give the host
free again when it is already overbooked so it
should recover the situation again.

That double-check should not hurt so we keep it
in just in case there are more scheduler bugs
somewhere.